+++ CHANGELOG.md
@@ -13,6 +13,11 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- Merchant Session resource
+- Merchant Installment resource
+- Merchant Purchase resource
+- Merchant Card resource
 
 ## [2.19.0] - 2025-03-17
 ### Added